added my sandbox, some clean-up
