+++ .gitignore
@@ -1,3 +1,7 @@
 Data\rsna-2023-abdominal-trauma-detection-minimal\segmentations\21057.nii
 model/*
 Data/*
+/data/*
+/model/*
+/venv/*
+/.idea/*
